% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animal_noises.R
\name{animal_noises}
\alias{animal_noises}
\title{Animal Noises}
\usage{
animal_noises(animal, noise)
}
\arguments{
\item{animal}{the name of the animal}

\item{noise}{the noise the animal makes}
}
\description{
Animal Noises
}
\examples{
animal_noises(animal = "cow", noise = "moo")
}
