% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salad_recipe.R
\name{salad_recipe}
\alias{salad_recipe}
\title{Salad Recipe}
\usage{
salad_recipe(ingredient1, lettuce_type, dressing)
}
\arguments{
\item{ingredient1}{name one ingredient in the salad}

\item{lettuce_type}{name the type of lettuce used in the salad}

\item{dressing}{name the type of dressing used in the salad}
}
\description{
Salad Recipe
}
\examples{
salad_recipe(ingredient1 = "carrot", lettuce_type = "iceberg", dressing = "miso")
}
