% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothie.R
\name{smoothie}
\alias{smoothie}
\title{Smoothie generator}
\usage{
smoothie(fruit, juice, enhancement)
}
\arguments{
\item{fruit}{name of the fruit to add into your smoothie}

\item{juice}{name of the juice to add into your smoothie}

\item{enhancement}{name of the enhancement to add to your smoothie}
}
\description{
Smoothie generator
}
\examples{
smoothie(fruit = "mango", juice = "guava", enhancement = "energy boost")
}
